Add e_lineoff entry to set the cursor width correctly
Decode LDR Rd,[PC,#offset] in e_plus to jump to the location referenced.

Pass the previous instruction to the disassembler when redrawing to make
sure it can decode BL (which is a two instruction sequence) properly
regardless of where redraw starts. This hasn't been a problem because the
Debugger module copes with the problem very gracefully.